backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
cohere==4.47
qdrant-client==1.7.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-multipart==0.0.6
httpx==0.26.0
